[update] : 設定ファイルの書き込み関連の関数を別ファイルへ
