+++ CMakeLists.txt
@@ -18,6 +18,7 @@ include_directories("inc")
 add_executable(alterlinux-i3-manager
   src/main.cpp
   src/mainwindow.cpp
+  src/mainwindow_looks_setting.cpp
   src/looks_page.cpp
 )
